Misc fix & improvement

+++ src/main/java/me/sjtumeow/meow/controller/api/web/AdminUserController.java
@@ -1,5 +1,7 @@
 package me.sjtumeow.meow.controller.api.web;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,9 +15,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import me.sjtumeow.meow.authorization.web.WebAuthUtility;
 import me.sjtumeow.meow.model.User;
-import me.sjtumeow.meow.model.form.AdminUpdateUserForm;
 import me.sjtumeow.meow.model.form.AdminRegisterForm;
+import me.sjtumeow.meow.model.form.AdminUpdateUserForm;
 import me.sjtumeow.meow.model.result.FailureMessageResult;
 import me.sjtumeow.meow.model.result.NewEntityIdResult;
 import me.sjtumeow.meow.service.UserService;
@@ -28,6 +31,9 @@ public class AdminUserController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private WebAuthUtility webAuthUtility;
 
     @GetMapping
     Iterable<User> getUsers(@RequestParam(required = false) Integer page, @RequestParam(required = false) Integer size,
@@ -61,18 +67,24 @@ public class AdminUserController {
     }
 
     @PatchMapping(path = "/{id}", consumes = "application/json")
-    ResponseEntity<?> updateUser(@PathVariable("id") Long id, @RequestBody AdminUpdateUserForm auuf) {
+    ResponseEntity<?> updateUser(@PathVariable("id") Long id, @RequestBody AdminUpdateUserForm auuf,
+            HttpSession session) {
 
         String password = auuf.getPassword();
         if (password != null && !FormatValidator.checkPassword(password))
             return ResponseEntity.badRequest().body(new FailureMessageResult("密码的长度至少为 6 个字符"));
+        if (webAuthUtility.getCurrentUser(session).getId().equals(id) && auuf.getIsAdmin() != null
+                && !auuf.getIsAdmin())
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new FailureMessageResult("禁止修改本人权限"));
 
         return userService.update(id, auuf) ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
 
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<?> deleteUser(@PathVariable("id") Long id) {
+    ResponseEntity<?> deleteUser(@PathVariable("id") Long id, HttpSession session) {
+        if (webAuthUtility.getCurrentUser(session).getId().equals(id))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new FailureMessageResult("禁止删除当前用户"));
         return userService.delete(id) ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
